refactor: move project portfolio to `data/projects`
